map::size() returns a size_t.
BUG=None
TEST=Compile with gcc.

Review URL: https://chromiumcodereview.appspot.com/23935006

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@221838 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5716b3ea630b391005509b9f255ab38c79761907

+++ cloud_print/gcp20/prototype/print_job_handler.cc
@@ -25,7 +25,7 @@ const int kErrorTimeoutSec = 30;
 // Errors simulation constants:
 const double kPaperJamProbability = 0.2;
 const int kTooManyDraftsTimeout = 10;
-const int kMaxDrafts = 5;
+const size_t kMaxDrafts = 5;
 
 const base::FilePath::CharType kJobsPath[] = FILE_PATH_LITERAL("printjobs");
